Update README.md
Event extraction is intended to extract from the text a characterization of an event, defined by a set of entities associated with a specific role in the event. aHistorical events are specific types of events that are framed in a historical context, reporting facts, dates, historical figures, and locations The novelty of this detecting historical events task in this paper lies in the following endeavors. firstly, an operative definition of the notion of historical events. Secondly, with the methodology, we try Focusing on specific types of events, specific people events, topic-related events, or short/long-term events. Thirdly, an overview and statistics about the events extracted from a historical corpus of interest. And a case study of so called “Mozart” which is the main propose of this project.